+++ src/cls/AdsPostsWindow.py
@@ -12,26 +12,37 @@ class AdsPostsWindow(tk.Frame):
         ## Top Frame
         topFrame = tk.Frame(self, pady=10)
 
+        # Username
+        userNameLabel = tk.Label(topFrame, text='Username')
+        userNameLabel.grid(row=0, sticky='E')
+
+        userNameEntry = tk.Entry(topFrame, textvariable=controller.userNameVar)
+        userNameEntry.grid(row=0, column=1, ipadx=15, padx=5)
+
+        # Blank
+        userNameLabel = tk.Label(topFrame)
+        userNameLabel.grid(row=1, sticky='E')
+
         # Email Row
-        emailLabel = tk.Label(topFrame, text='Email')
-        emailLabel.grid(row=0, sticky='E')
+        emailLabel = tk.Label(topFrame, text='Facebook Email/Phone')
+        emailLabel.grid(row=2, sticky='E')
 
         emailEntry = tk.Entry(topFrame, textvariable=controller.emailVar)
-        emailEntry.grid(row=0, column=1, ipadx=15, padx=5)
+        emailEntry.grid(row=2, column=1, ipadx=15, padx=5)
 
         # Password Row
-        passLabel = tk.Label(topFrame, text='Password')
-        passLabel.grid(row=1, sticky='E')
+        passLabel = tk.Label(topFrame, text='Facebook Password')
+        passLabel.grid(row=3, sticky='E')
 
         passEntry = tk.Entry(topFrame, textvariable=controller.passVar, show='*')
-        passEntry.grid(row=1, column=1, ipadx=15)
+        passEntry.grid(row=3, column=1, ipadx=15)
 
         # Telegram ID Row
         teleIdLabel = tk.Label(topFrame, text='Telegram User ID')
-        teleIdLabel.grid(row=2, sticky='E')
+        teleIdLabel.grid(row=4, sticky='E')
 
         teleIdEntry = tk.Entry(topFrame, textvariable=controller.teleIdVar)
-        teleIdEntry.grid(row=2, column=1, ipadx=15)
+        teleIdEntry.grid(row=4, column=1, ipadx=15)
 
         # Remember Me Checkbox
         rememberMeCB = tk.Checkbutton(topFrame, text='Remember Me', variable=controller.rememberMeVar)
@@ -39,17 +50,17 @@ class AdsPostsWindow(tk.Frame):
 
         # Keywords Row
         keywordsLabel = tk.Label(topFrame, text='Keywords')
-        keywordsLabel.grid(row=4, sticky='E')
+        keywordsLabel.grid(row=6, sticky='E')
 
         keywordsEntry = tk.Entry(topFrame, textvariable=controller.keywordsVar)
-        keywordsEntry.grid(row=4, column=1, ipadx=15)
+        keywordsEntry.grid(row=6, column=1, ipadx=15)
 
         # Blacklist Keywords Row
         blacklistKeywordsLabel = tk.Label(topFrame, text='Blacklist Keywords')
-        blacklistKeywordsLabel.grid(row=5, sticky='E')
+        blacklistKeywordsLabel.grid(row=7, sticky='E')
 
         blacklistKeywordsEntry = tk.Entry(topFrame, textvariable=controller.blacklistKeywordsVar)
-        blacklistKeywordsEntry.grid(row=5, column=1, ipadx=15)
+        blacklistKeywordsEntry.grid(row=7, column=1, ipadx=15)
 
         topFrame.pack()
         ## End Top Frame
@@ -72,7 +83,7 @@ class AdsPostsWindow(tk.Frame):
         lowerFrame.pack()
 
     def start_scrape_ads_thread(self):
-        scrapeAdsThread = threading.Thread(target=scrape_ads, args=(controller.version, controller.statusBar, controller.chromePath,
-                                        controller.session_id, controller.keywordsVar, controller.blacklistKeywordsVar, controller.emailVar,
-                                        controller.passVar, controller.teleIdVar, controller.oldUsersList,), daemon=True, name='scraping_ads_thread')
+        scrapeAdsThread = threading.Thread(target=scrape_ads, args=(self.controller.userNameVar, self.controller.version, self.controller.statusBar, self.controller.chromePath,
+                                        self.controller.session_id, self.controller.keywordsVar, self.controller.blacklistKeywordsVar, self.controller.emailVar,
+                                        self.controller.passVar, self.controller.teleIdVar, self.controller.oldUsersList,), daemon=True, name='scraping_ads_thread')
         scrapeAdsThread.start()