+++ src/cls/GroupPostsWindow.py
@@ -12,26 +12,37 @@ class GroupPostsWindow(tk.Frame):
         ## Top Frame
         topFrame = tk.Frame(self, pady=5)
 
+        # Username
+        userNameLabel = tk.Label(topFrame, text='Username')
+        userNameLabel.grid(row=0, sticky='E')
+
+        userNameEntry = tk.Entry(topFrame, textvariable=controller.userNameVar)
+        userNameEntry.grid(row=0, column=1, ipadx=15, padx=5)
+
+        # Blank
+        userNameLabel = tk.Label(topFrame)
+        userNameLabel.grid(row=1, sticky='E')
+
         # Email Row
         emailLabel = tk.Label(topFrame, text='Email')
-        emailLabel.grid(row=0, sticky='E')
+        emailLabel.grid(row=2, sticky='E')
 
         emailEntry = tk.Entry(topFrame, textvariable=controller.emailVar2)
-        emailEntry.grid(row=0, column=1, ipadx=15, padx=5)
+        emailEntry.grid(row=2, column=1, ipadx=15, padx=5)
 
         # Password Row
         passLabel = tk.Label(topFrame, text='Password')
-        passLabel.grid(row=1, sticky='E')
+        passLabel.grid(row=3, sticky='E')
 
         passEntry = tk.Entry(topFrame, textvariable=controller.passVar2, show='*')
-        passEntry.grid(row=1, column=1, ipadx=15)
+        passEntry.grid(row=3, column=1, ipadx=15)
 
         # Telegram ID Row
         teleIdLabel = tk.Label(topFrame, text='Telegram User ID')
-        teleIdLabel.grid(row=2, sticky='E')
+        teleIdLabel.grid(row=4, sticky='E')
 
         teleIdEntry = tk.Entry(topFrame, textvariable=controller.teleIdVar2)
-        teleIdEntry.grid(row=2, column=1, ipadx=15)
+        teleIdEntry.grid(row=4, column=1, ipadx=15)
 
         # Remember Me Checkbox
         rememberMeCB = tk.Checkbutton(topFrame, text='Remember Me', variable=controller.rememberMeVar2)
@@ -39,24 +50,24 @@ class GroupPostsWindow(tk.Frame):
 
         # Keywords Row
         keywordsLabel = tk.Label(topFrame, text='Keywords')
-        keywordsLabel.grid(row=4, sticky='E')
+        keywordsLabel.grid(row=6, sticky='E')
 
         keywordsEntry = tk.Entry(topFrame, textvariable=controller.keywordsVar2)
-        keywordsEntry.grid(row=4, column=1, ipadx=15)
+        keywordsEntry.grid(row=6, column=1, ipadx=15)
 
         # Blacklist Keywords Row
         blacklistKeywordsLabel = tk.Label(topFrame, text='Blacklist Keywords')
-        blacklistKeywordsLabel.grid(row=5, sticky='E')
+        blacklistKeywordsLabel.grid(row=7, sticky='E')
 
         blacklistKeywordsEntry = tk.Entry(topFrame, textvariable=controller.blacklistKeywordsVar2)
-        blacklistKeywordsEntry.grid(row=5, column=1, ipadx=15)
+        blacklistKeywordsEntry.grid(row=7, column=1, ipadx=15)
 
         # Group ID List
         groupIdListLabel = tk.Label(topFrame, text='Group IDs')
-        groupIdListLabel.grid(row=6, sticky='E')
+        groupIdListLabel.grid(row=8, sticky='E')
 
         groupIdListEntry = tk.Entry(topFrame, textvariable=controller.groupIdListVar)
-        groupIdListEntry.grid(row=6, column=1, ipadx=15)
+        groupIdListEntry.grid(row=8, column=1, ipadx=15)
 
         topFrame.pack()
         ## End Top Frame
@@ -79,7 +90,7 @@ class GroupPostsWindow(tk.Frame):
         lowerFrame.pack()
 
     def start_scrape_groups_thread(self):
-        scrapeGroupsThread = threading.Thread(target=scrape_groups, args=(self.controller.groupIdListVar, self.controller.version, self.controller.statusBar,
+        scrapeGroupsThread = threading.Thread(target=scrape_groups, args=(self.controller.userNameVar, self.controller.groupIdListVar, self.controller.version, self.controller.statusBar,
                                         self.controller.chromePath, self.controller.session_id, self.controller.keywordsVar2, self.controller.blacklistKeywordsVar2,
                                         self.controller.emailVar2, self.controller.passVar2, self.controller.teleIdVar2, self.controller.oldUsersList,),
                                         daemon=True, name='scraping_groups_thread')