+++ src/sample-accounts.py
@@ -1,0 +1,14 @@
+"""
+This is the sample of module accounts.py which needed for the application to function normally.
+Fill all the token below and save as accounts.py.
+
+Beside of this file, you will need your token of Google Accounts for OAuth, save it as token.pickle and store in this dir.
+Check https://developers.google.com/sheets/api/quickstart/python for details.
+"""
+
+botToken = ''
+spreadsheetIdHubspot = ''
+spreadsheetIdData = ''
+spreadsheetIdInfo = ''
+spreadsheetIdPosts = ''
+spreadsheetIdGroupPosts = ''