+++ src/utils/checkValidation.py
@@ -16,11 +16,11 @@ def check_validation(type, version=None, email=None, teleId=None):
     else:
         if type == 'version':
             if checkDf.active.iloc[0] == 'FALSE':
-                if checkDf.message.iloc[0] == 'deprecated':
-                    messagebox.showerror(title='Version Deprecated', message='This version has been deprecated.\nPlease download the latest version.')
-                    quit()
-                elif checkDf.message.iloc[0] == 'maintain':
+                if checkDf.message.iloc[0] == 'maintain':
                     messagebox.showerror(title='Maintain', message='The app is under maintainance. Please try again later.')
+                    quit()
+                else:
+                    messagebox.showerror(title='Version Deprecated', message='This version has been deprecated.\nPlease download the latest version.')
                     quit()
 
         elif type == 'user':